Attach additional secgroup for inter instance traffic
When using the "Classic Provider" network name, some security group rules
which enable the internal traffic among each instance cannot be generated.
In order to allow This, this patch will generate a common internal
self-reference secgroup which is attached to every instance of the cluster.

The patch could be merged with #commit 4ba70f4eac794084501907787f5ace345f990af2

Change-Id: I2ecf0a766cc95750f4653c2a72b346bc28f1bab2